CRUD PRODUTOS E CODIGO DE BARRAS
feat(produtos): CRUD completo de produtos, porém o incremento de produtos no estoque ainda não está funcionando.
feat(barras): Geração do código de barras funcionando

+++ Projeto/painel-adm/categorias/excluir.php
@@ -6,6 +6,16 @@ $id = $_POST['id'];
 //BUSCAR IMAGEM PARA EXLUSÃO NA PASTA
 $query_con = $pdo->query("SELECT * FROM categorias WHERE id = '$id'");
 $res_con = $query_con->fetchAll(pdo::FETCH_ASSOC);
+
+//VERIFICAR SE EXISTE REGISTRO RELACIONADO
+$query_p = $pdo->query("SELECT * FROM produtos WHERE categoria = '$id'");
+$res_p = $query_p->fetchAll(pdo::FETCH_ASSOC);
+if(@count($res_p) > 0){
+    echo 'Existem produtos cadastrados com essa categoria, exclua o mesmo para depois excluir a categoria';
+    exit();
+}
+
+//EXCLUIR IMAGEM DA PASTA
 $imagem = $res_con[0]['foto'];
 unlink('../../img/categorias/' .$imagem);
